Move FindRegions to contrib
